requirements_new.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
dotenv==0.9.9
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
jwcrypto==1.5.5
Markdown==3.7
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pytest==8.1.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.0
responses==0.23.3
sqlparse==0.5.3
tomli==2.2.1
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
